pyproject.toml
[project]
name = "rag-invenio-ai"
version = "0.1.0"
description = "Advanced RAG Pipeline with Groq, FastEmbed, and Streamlit"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "streamlit>=1.31.0",
    "requests>=2.32.3",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.1",
    "langchain>=0.3.7",
    "langchain-core>=0.3.17",
    "langchain-community>=0.3.7",
    "langchain-groq>=0.2.1",
    "langchain-qdrant>=0.2.0",
    "langchain-classic>=1.0.5",
    "langchain-experimental>=0.3.4",
    "qdrant-client>=1.12.1",
    "pymupdf>=1.25.1",
    "fastembed>=0.5.1",
    "flashrank>=0.2.8",
    "rank-bm25>=0.2.2",
    "diskcache>=5.6.3",
    "redis>=5.2.1",
    "pandas>=2.0.0",
    "plotly>=5.18.0",
    "tokenizers==0.22.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "langchain-text-splitters>=1.1.2",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["backend/tests"]
pythonpath = ["backend"]

